Final fixups

+++ MP1/test/server.sh
@@ -2,6 +2,11 @@
 
 make -C ../src &> /dev/null
 
+if [ $? -ne 0 ]; then
+    echo "Make failed"
+    exit 1
+fi
+
 rm -f users.txt *.following *.timeline
 
 ../src/tsd
